Move query generation code
